✅ test(tui): the PATH wiring test plants its own home instead of borrowing the real one
Codex review, third pass. Two findings on the render test added in the
previous commit, both real, and one fix answers both.

**Separator.** The assertion built its needle from `MAIN_SEPARATOR`, which
is `\` on Windows, while the rendered value keeps the separator libgit2
emits, which is `/` there. It would have gone red on `windows-latest` only,
the exact failure mode this repo has been bitten by before.

**Ambient state.** `TempDir::new_in(dirs::home_dir())` wrote into the user's
real home and panicked for an environmental reason if that home were absent
or read-only, which AGENTS.md calls out by name.

Both go away by planting the home rather than borrowing it: the test moves
to its own binary and sets `HOME` to a `TempDir` before anything reads one.
Its own binary is the mechanism, not tidiness — the lookup sits behind a
`OnceLock`, so the first `dirs::home_dir()` anywhere in a binary freezes the
value for every later test in it, and only a single-test binary makes the
injection deterministic. `#[cfg(unix)]` because the injection is: Windows
resolves the home through `SHGetKnownFolderPath` and ignores the variable.
The rule the compression applies there is covered by the unit test, which
needs no fixture, and the wiring under test is not platform-dependent.

Two things the first draft got wrong and the second does not. The planted
home is canonicalised, because a real `$HOME` is and a `TempDir` is not: on
macOS `/var` is a symlink to `/private/var`, libgit2 resolves it and the
environment does not, so the raw temp path never prefix-matched the rendered
one. And the test takes the binary's env lock, because
`env_guard_invariant_tests` derives its audit from the `set_var` calls
themselves and grants no exemptions — the lock serialises nothing with one
test in the file, and exists so a second one inherits it rather than
discovering it.

Still proven red by reverting the one line it guards, and pre-validated
against a stripped environment as AGENTS.md asks: green under
`env -u HOME PATH="$(dirname "$(command -v cargo)"):/usr/bin:/bin"`.

refs #568
